adding handler func and updating repo, model and tests

+++ internal/repository/products/proto.go
@@ -8,7 +8,7 @@ import (
 
 // ProductResponse struct
 type ProductResponse struct {
-	ID          int                  `json:"id_product" gorm:"primaryKey"`
+	ID          uint64               `json:"id_product" gorm:"primaryKey"`
 	Name        string               `json:"name" gorm:"not null;unique"`
 	Description string               `json:"description"`
 	Price       float64              `json:"price" gorm:"not null;check:price >= 0"`
