+++ model/products/product_history.go
@@ -13,8 +13,8 @@ const (
 )
 
 type ProductHistory struct {
-	ID          int           `json:"id_product_history" gorm:"primaryKey"`
-	IDProduct   int           `json:"id_product" gorm:"primaryKey"`
+	ID          uint64        `json:"id_product_history" gorm:"primaryKey"`
+	IDProduct   uint64        `json:"id_product" gorm:"primaryKey"`
 	Name        string        `json:"name" gorm:"not null;unique"`
 	Description string        `json:"description"`
 	Price       float64       `json:"price" gorm:"not null;check:price >= 0"`
